chore: refactors session manager class.
